refactor: merge fodlers
